-Added reminder to set up their MongoDB connection -Simplified league controller -Combined single use interfaces into LeaguesRepository and MongoDbContext -Commented out  unnecessary NavLinks from NavMenu
