+++ Services/LeaguesRepository.cs
@@ -7,6 +7,15 @@ using TagProLeague.Models;
 
 namespace TagProLeague.Services
 {
+    public interface ILeaguesRepository
+    {
+        Task<IEnumerable<League>> GetAllLeagues();
+        Task<League> GetLeague(string name);
+        Task CreateLeague(League league);
+        Task<bool> UpdateLeague(League league);
+        Task<bool> DeleteLeague(string name);
+    }
+
     public class LeaguesRepository : ILeaguesRepository
     {
         private readonly IMongoDbContext _context;
@@ -15,6 +24,7 @@ namespace TagProLeague.Services
         {
             _context = context;
         }
+
         public async Task<IEnumerable<League>> GetAllLeagues()
         {
             return await _context
@@ -22,6 +32,7 @@ namespace TagProLeague.Services
                             .Find(_ => true)
                             .ToListAsync();
         }
+
         public Task<League> GetLeague(string name)
         {
             FilterDefinition<League> filter = Builders<League>.Filter.Eq(m => m.Name, name);
@@ -35,6 +46,7 @@ namespace TagProLeague.Services
         {
             await _context.Leagues.InsertOneAsync(league);
         }
+
         public async Task<bool> UpdateLeague(League league)
         {
             ReplaceOneResult updateResult =
@@ -46,6 +58,7 @@ namespace TagProLeague.Services
             return updateResult.IsAcknowledged
                     && updateResult.ModifiedCount > 0;
         }
+
         public async Task<bool> DeleteLeague(string name)
         {
             FilterDefinition<League> filter = Builders<League>.Filter.Eq(m => m.Name, name);
