+++ appsettings.json
@@ -11,5 +11,7 @@
         "Default": "Warning"
       }
     }
-  }
+  },
+  "MongoDb:ConnectionString": "Set in User Secrets",
+  "MongoDb:Database": "Set in User Secrets"
 }
